style:스터디 css 수정
