+++ src/styles/components/_todoList.scss
@@ -1,0 +1,58 @@
+.todo-container {
+  border: 1px solid $study-border;
+  border-radius: 6px;
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  @include flexbox();
+
+  p {
+    @include text-style-24;
+    padding: 10px;
+    color: $black;
+  }
+
+  .todo-content {
+    flex: 1;
+    width: 100%;
+    display: flex;
+    @include flexbox();
+    padding: 10px;
+
+    input {
+      border: 1px solid$study-border;
+      border-radius: 6px;
+      width: 100%;
+      height: 50px;
+      text-indent: 10px;
+    }
+  }
+}
+
+.todo-content-container {
+  width: 100%;
+  height: 100%;
+}
+
+.edit-todo-task {
+  margin-bottom: 20px;
+}
+
+.todo-task {
+  border: 1px solid $study-border;
+  border-radius: 6px;
+  padding: 10px;
+  display: flex;
+  @include flexbox(space-between, center);
+  height: 50px;
+  margin-bottom: 20px;
+}
+
+.todo-form-container {
+  display: flex;
+  flex-direction: column;
+  @include flexbox();
+  gap: 30px;
+  width: 100%;
+  height: 100%;
+}
