+++ src/styles/main.scss
@@ -32,6 +32,7 @@
 @import 'components/studyScedule';
 @import 'components/commentContainer';
 @import 'components/calendar';
+@import 'components/todoList';
 
 /* 레이아웃 */
 @import 'layouts/grids';
